Prefer direct node connections in certain cases
That is, if node_b is set to SIZE_SHRINK_BEGIN and node_a is in an even
position, or node_b is set to SIZE_SHRINK_END and node_a is odd.

Doing so might result in diagonal lines under certain conditions¹, but
avoids intersecting lines, which I consider a worthy compromise.

¹ Mostly when the size flags used on node_a "counteract" on those used
on node_b. For example, SIZE_SHRINK_BEGIN on node_b and on node_a will
work fine, but if node_a is changed to SIZE_SHRINK_END and the container
is set to expand and fill, that will result in a diagonal connection.